Refactor classes into their respective files
